First pass at refactor.
